requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
google-genai>=0.3.0
python-dotenv>=1.0.0
